Support prefixed custom_values method on the Liquid drops
Use the attribute name as the prefix when fetching the value of the custom field that's not on the model the drop is for.
For example, to get the signup custom values on a donation page, this is the Liquid variable one has to use:

signup.custom_values.<custom_field_name>, where

<custom_field_name> is the slug of an existing signup custom field. To get this custom field's value from the drop,
a method called signup_custom_values has to be defined in the drop.